+++ winery-data/src/main/java/com/changfa/frame/data/repository/user/AdminUserRepository.java
@@ -17,28 +17,28 @@ public interface AdminUserRepository extends AdvancedJpaRepository<AdminUser,Int
     AdminUser findAdminUserByUserName(String userName,Integer wid);
 
     @Query(value = "select * from admin_user where phone = ?1 and winery_id=?2 ",nativeQuery = true)
-    AdminUser findAdminUserByPhone(String phone,Integer wineryId);
+    AdminUser findAdminUserByPhone(String phone,Long wineryId);
 
     @Query(value = "select * from admin_user where phone = ?1 ",nativeQuery = true)
     AdminUser findAdminUserByPhone(String phone);
 
     @Query(value = "SELECT DISTINCT au.* from admin_user au,user_role ur,role r where au.id = ur.user_id and ur.role_id = r.id and au.winery_id = ?1 and (au.user_name like CONCAT('%',?2,'%') or au.phone like CONCAT('%',?2,'%') or r.name like CONCAT('%',?2,'%')) order by id desc",nativeQuery = true)
-    List<AdminUser> findAdminUserByWId(int wineryId, String search);
+    List<AdminUser> findAdminUserByWId(Long wineryId, String search);
 
     @Query(value = "SELECT au.* \n" +
             "FROM admin_user au \n" +
             "LEFT JOIN user_role ur ON au.id = ur.user_id\n" +
             "LEFT JOIN role r ON ur.role_id = r.id\n" +
             "WHERE au.winery_id = ?1 AND r.name = ?2 AND au.status = 'A'",nativeQuery = true)
-    List<AdminUser> findAdminUserByWIdAndRoleName(int wineryId, String roleName);
+    List<AdminUser> findAdminUserByWIdAndRoleName(Long wineryId, String roleName);
 
     //获取酒庄的超级管理员用户
     @Query("select a from AdminUser a where a.wineryId = ?1 and a.isSuper = 'Y' ")
-    AdminUser findAdminUserByWineryId(Integer wineryId);
+    AdminUser findAdminUserByWineryId(Long wineryId);
 
 
     /*酒庄下的工作人员数量*/
-    int countAdminUsersByWineryIdAndStatus(Integer wineryId,  String status);
+    int countAdminUsersByWineryIdAndStatus(Long wineryId,  String status);
